pyproject.toml
[project]
name = "news-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "dash>=3.0.4",
    "datasets>=3.6.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "vadersentiment>=3.3.2",
]
